# Use github.com/mailru/easyjson for JSON encoding and decoding of Targets and Results