--- a/Gopkg.toml
+++ b/Gopkg.toml
@@ -68,3 +68,7 @@
 [[constraint]]
   branch = "master"
   name = "github.com/c2h5oh/datasize"
+
+[[constraint]]
+  branch = "master"
+  name = "github.com/mailru/easyjson"