--- a/lib/results_test.go
+++ b/lib/results_test.go
@@ -3,6 +3,7 @@ package vegeta
 import (
 	"bytes"
 	"io"
+	"math/rand"
 	"reflect"
 	"testing"
 	"testing/quick"
@@ -79,19 +80,23 @@ func TestEncoding(t *testing.T) {
 
 				var buf bytes.Buffer
 				enc := tc.enc(&buf)
-				if err := enc(&want); err != nil {
-					t.Fatal(err)
+				for j := 0; j < 2; j++ {
+					if err := enc(&want); err != nil {
+						t.Fatal(err)
+					}
 				}
 
 				dec := tc.dec(&buf)
-				var got Result
-				if err := dec(&got); err != nil {
-					t.Fatalf("err: %q buffer: %s", err, buf.String())
-				}
+				for j := 0; j < 2; j++ {
+					var got Result
+					if err := dec(&got); err != nil {
+						t.Fatalf("err: %q buffer: %s", err, buf.String())
+					}
 
-				if !got.Equal(want) {
-					t.Logf("\ngot:  %#v\nwant: %#v\n", got, want)
-					return false
+					if !got.Equal(want) {
+						t.Logf("\ngot:  %#v\nwant: %#v\n", got, want)
+						return false
+					}
 				}
 
 				return true
@@ -99,6 +104,51 @@ func TestEncoding(t *testing.T) {
 
 			if err != nil {
 				t.Fatal(err)
+			}
+		})
+	}
+}
+
+func BenchmarkEncodings(b *testing.B) {
+	b.StopTimer()
+	b.ResetTimer()
+
+	rng := rand.New(rand.NewSource(0))
+	zf := rand.NewZipf(rng, 3, 2, 1000)
+	began := time.Now()
+	results := make([]Result, 1e5)
+
+	for i := 0; i < cap(results); i++ {
+		results[i] = Result{
+			Attack:    "Big Bang!",
+			Seq:       uint64(i),
+			Timestamp: began.Add(time.Duration(i) * time.Millisecond),
+			Latency:   time.Duration(zf.Uint64()) * time.Millisecond,
+		}
+	}
+
+	for _, tc := range []struct {
+		encoding string
+		enc      func(io.Writer) Encoder
+		dec      func(io.Reader) Decoder
+	}{
+		{"gob", NewEncoder, NewDecoder},
+		{"csv", NewCSVEncoder, NewCSVDecoder},
+		{"json", NewJSONEncoder, NewJSONDecoder},
+	} {
+		var buf bytes.Buffer
+		enc := tc.enc(&buf)
+
+		b.Run(tc.encoding+"-encode", func(b *testing.B) {
+			for i := 0; i < b.N; i++ {
+				enc.Encode(&results[i%len(results)])
+			}
+		})
+
+		dec := tc.dec(&buf)
+		b.Run(tc.encoding+"-decode", func(b *testing.B) {
+			for i := 0; i < b.N; i++ {
+				dec.Decode(&results[i%len(results)])
 			}
 		})
 	}