It's much faster.

```
benchmark                            old ns/op     new ns/op     delta
BenchmarkEncodings/gob-encode-8      1278          1146          -10.33%
BenchmarkEncodings/gob-decode-8      131           128           -2.29%
BenchmarkEncodings/csv-encode-8      800           757           -5.38%
BenchmarkEncodings/csv-decode-8      705           657           -6.81%
BenchmarkEncodings/json-encode-8     2617          1212          -53.69%
BenchmarkEncodings/json-decode-8     6259          1547          -75.28%

benchmark                            old allocs     new allocs     delta
BenchmarkEncodings/gob-encode-8      2              2              +0.00%
BenchmarkEncodings/gob-decode-8      0              0              +0.00%
BenchmarkEncodings/csv-encode-8      2              2              +0.00%
BenchmarkEncodings/csv-decode-8      2              2              +0.00%
BenchmarkEncodings/json-encode-8     3              4              +33.33%
BenchmarkEncodings/json-decode-8     7              2              -71.43%

benchmark                            old bytes     new bytes     delta
BenchmarkEncodings/gob-encode-8      200           200           +0.00%
BenchmarkEncodings/gob-decode-8      0             0             +0.00%
BenchmarkEncodings/csv-encode-8      209           209           +0.00%
BenchmarkEncodings/csv-decode-8      192           192           +0.00%
BenchmarkEncodings/json-encode-8     801           1004          +25.34%
BenchmarkEncodings/json-decode-8     72            64            -11.11%
```

```
benchmark                                old ns/op     new ns/op     delta
BenchmarkJSONTargetEncoding/encode-8     787           781           -0.76%
BenchmarkJSONTargetEncoding/decode-8     5812          1510          -74.02%

benchmark                                old allocs     new allocs     delta
BenchmarkJSONTargetEncoding/encode-8     1              1              +0.00%
BenchmarkJSONTargetEncoding/decode-8     18             9              -50.00%

benchmark                                old bytes     new bytes     delta
BenchmarkJSONTargetEncoding/encode-8     470           470           +0.00%
BenchmarkJSONTargetEncoding/decode-8     1230          703           -42.85%
```